SOL-561 merged bundles tests to products tests.
