+++ cypress/e2e/api/products/products.cy.ts
@@ -34,6 +34,7 @@ describe('products backend api', { tags: ['@api', '@products', 'product'] }, ():
   let accessToken: string;
   let sku: string;
   let abstractSku: string;
+  let bundledSku: string;
 
   // Builds a minimal valid POST body — every referenced natural key exists (Policy B enforces 422 otherwise).
   const buildValidProductBody = (concreteSku: string, parentAbstractSku: string | null): Record<string, unknown> => ({
@@ -60,6 +61,7 @@ describe('products backend api', { tags: ['@api', '@products', 'product'] }, ():
     ({ dynamicFixtures, staticFixtures } = Cypress.env());
     sku = dynamicFixtures.product.sku;
     abstractSku = dynamicFixtures.product.abstract_sku;
+    bundledSku = dynamicFixtures.bundled.sku;
   });
 
   beforeEach((): void => {
@@ -104,9 +106,13 @@ describe('products backend api', { tags: ['@api', '@products', 'product'] }, ():
         expect(attributes.imageSets, 'imageSets').to.be.an('array').and.to.have.length.greaterThan(0);
         expect(attributes.imageSets[0].images, 'images').to.have.length.greaterThan(0);
 
-        expect(attributes.localizedAttributes, 'localizedAttributes').to.be.an('array').and.to.have.length.greaterThan(0);
+        expect(attributes.localizedAttributes, 'localizedAttributes')
+          .to.be.an('array')
+          .and.to.have.length.greaterThan(0);
         expect(
-          attributes.localizedAttributes.find((entry) => entry.localeName === staticFixtures.localeName),
+          attributes.localizedAttributes.find(
+            (entry: Record<string, unknown>) => entry.localeName === staticFixtures.localeName
+          ),
           'seeded locale present'
         ).to.exist;
 
@@ -330,14 +336,12 @@ describe('products backend api', { tags: ['@api', '@products', 'product'] }, ():
 
     it('should update localized attributes', (): void => {
       updateProduct(accessToken, patchSku, {
-        localizedAttributes: [
-          { localeName: staticFixtures.localeName, name: 'Patched name', isSearchable: false },
-        ],
+        localizedAttributes: [{ localeName: staticFixtures.localeName, name: 'Patched name', isSearchable: false }],
       }).then((response) => {
         expect(response.status).to.eq(200);
         expect(
           response.body.data.attributes.localizedAttributes.find(
-            (entry) => entry.localeName === staticFixtures.localeName
+            (entry: Record<string, unknown>) => entry.localeName === staticFixtures.localeName
           ).name
         ).to.eq('Patched name');
       });
@@ -850,6 +854,76 @@ describe('products backend api', { tags: ['@api', '@products', 'product'] }, ():
         failOnStatusCode: false,
       }).then((response) => {
         expect(response.status).to.eq(400);
+      });
+    });
+  });
+
+  describe('bundles', (): void => {
+    // Bundle products carry no own stock rows — availability derives from their components,
+    // so the bundle is created without a `stocks` field (unlike a standalone concrete).
+    const buildBundleBody = (bundleSku: string): Record<string, unknown> => ({
+      sku: bundleSku,
+      isActive: true,
+      attributes: { color: 'blue' },
+      localizedAttributes: [{ localeName: staticFixtures.localeName, name: `Bundle ${bundleSku}`, isSearchable: true }],
+      prices: [
+        {
+          priceTypeName: staticFixtures.priceTypeName,
+          storeName: staticFixtures.storeName,
+          currencyCode: staticFixtures.currencyCode,
+          netAmount: staticFixtures.netAmount,
+          grossAmount: staticFixtures.grossAmount,
+        },
+      ],
+      productBundle: [{ sku: bundledSku, quantity: staticFixtures.bundleQuantity }],
+    });
+
+    it('should expose the seeded stock on the standalone bundled product', (): void => {
+      getProduct(accessToken, bundledSku).then((response) => {
+        expect(response.status).to.eq(200);
+
+        const stock = response.body.data.attributes.stocks.find(
+          (item: Record<string, unknown>) => item.stockName === staticFixtures.stockName
+        );
+        expect(stock, 'bundled stock present').to.not.be.undefined;
+        expect(stock.quantity).to.eq(staticFixtures.bundledStockQuantity);
+        expect(response.body.data.attributes.productBundle, 'component is not itself a bundle').to.be.empty;
+      });
+    });
+
+    it('should assign a bundled product when creating a bundle', (): void => {
+      const bundleSku = `pxm-bundle-${Date.now()}`;
+
+      createProduct(accessToken, buildBundleBody(bundleSku)).then((response) => {
+        expect(response.status).to.be.oneOf([200, 201]);
+
+        const bundle = response.body.data.attributes.productBundle.find(
+          (item: Record<string, unknown>) => item.sku === bundledSku
+        );
+        expect(bundle, 'bundled product assigned').to.not.be.undefined;
+        expect(bundle.quantity).to.eq(staticFixtures.bundleQuantity);
+      });
+    });
+
+    it('should persist the bundle assignment on a subsequent read', (): void => {
+      const bundleSku = `pxm-bundle-read-${Date.now()}`;
+
+      createProduct(accessToken, buildBundleBody(bundleSku)).then((createResponse) => {
+        expect(createResponse.status, 'bundle created').to.be.oneOf([200, 201]);
+
+        getProduct(accessToken, bundleSku).then((response) => {
+          expect(response.status).to.eq(200);
+
+          const bundle = response.body.data.attributes.productBundle.find(
+            (item: Record<string, unknown>) => item.sku === bundledSku
+          );
+          expect(bundle, 'bundled product persisted').to.not.be.undefined;
+          expect(bundle.quantity).to.eq(staticFixtures.bundleQuantity);
+
+          // Bundle stock behaviour differs from a standalone concrete — logged for observation
+          // rather than asserted blind; tighten once the real shape is confirmed.
+          cy.log(`bundle stocks: ${JSON.stringify(response.body.data.attributes.stocks)}`);
+        });
       });
     });
   });
