+++ cypress/support/types/api/products.ts
@@ -9,10 +9,18 @@ export interface ProductsStaticFixtures {
   concreteAttributeValue: string;
   netAmount: number;
   grossAmount: number;
+  bundledStockQuantity: number;
+  bundleQuantity: number;
 }
 
 export interface ProductsDynamicFixtures {
   product: {
+    abstract_sku: string;
+    sku: string;
+    id_product_concrete: number;
+    fk_product_abstract: number;
+  };
+  bundled: {
     abstract_sku: string;
     sku: string;
     id_product_concrete: number;
@@ -28,26 +36,5 @@ export interface ProductsDynamicFixtures {
   };
   shipmentType: {
     uuid: string;
-  };
-}
-
-export interface ProductsBundleStaticFixtures {
-  storeName: string;
-  currencyCode: string;
-  priceTypeName: string;
-  localeName: string;
-  stockName: string;
-  bundledStockQuantity: number;
-  bundleQuantity: number;
-  netAmount: number;
-  grossAmount: number;
-}
-
-export interface ProductsBundleDynamicFixtures {
-  bundled: {
-    abstract_sku: string;
-    sku: string;
-    id_product_concrete: number;
-    fk_product_abstract: number;
   };
 }
